webSurvey

+++ WebSurvey_Sales_CRM/Areas/Admin/Reponsitory/Service/NewOptionService.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Http.HttpResults;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using WebSurvey_Sales_CRM.Areas.Admin.Reponsitory.Interface;
 using WebSurvey_Sales_CRM.Data;
@@ -11,9 +12,9 @@ namespace WebSurvey_Sales_CRM.Areas.Admin.Reponsitory.Service
         public readonly ApplicationDbContext _context;
         public NewOptionService(ApplicationDbContext context)
         {
-            this._context = context;
+            _context = context;
         }
-
+/*-------------------------------------------------- Roles --------------------------------------------------*/
         //Get all name roles in RolesUser
         [HttpGet]
         public async Task<IEnumerable<RolesUser>> GetRoles()
@@ -35,6 +36,11 @@ namespace WebSurvey_Sales_CRM.Areas.Admin.Reponsitory.Service
         {
             try
             {
+                if (rolesUser == null)
+                {
+                    return Enumerable.Empty<RolesUser>();
+                }
+
                 await _context.RolesUsers.AddAsync(rolesUser);
                 await _context.SaveChangesAsync();
                 return _context.RolesUsers;
@@ -45,7 +51,7 @@ namespace WebSurvey_Sales_CRM.Areas.Admin.Reponsitory.Service
                 throw;
             }
         }
-        //index delete
+        //find row data by id 
         [HttpGet]
         public async Task<IEnumerable<RolesUser>> GetDeleteNameRoles(int id)
         {
@@ -60,14 +66,17 @@ namespace WebSurvey_Sales_CRM.Areas.Admin.Reponsitory.Service
                 throw;
             }
         }
-
-        //delete
+        //delete name roles
         [HttpPost]
         public async Task<IEnumerable<RolesUser>> DeleteNameRoles(int id)
         {
             try
             {
                 var data = await _context.RolesUsers.FindAsync(id);
+                if(data == null)
+                {
+                    return Enumerable.Empty<RolesUser>();
+                }    
                 _context.RolesUsers.Remove(data);
                 _context.SaveChanges();
                 return _context.RolesUsers;
@@ -79,9 +88,78 @@ namespace WebSurvey_Sales_CRM.Areas.Admin.Reponsitory.Service
             }
         }
 
+/*-------------------------------------------------- Source --------------------------------------------------*/
+        //Get all name Source in Source
+        [HttpGet]
+        public async Task<IEnumerable<Source>> GetSource()
+        {
+            try
+            {
+                return await _context.Sources.ToListAsync();
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Error occurred: {ex.Message}");
+                throw;
+            }
 
+        }
+        //Add name roles in RolesUser
+        [HttpPost]
+        public async Task<IEnumerable<Source>> AddNameSource(Source source)
+        {
+            try
+            {
+                if (source == null)
+                {
+                    return Enumerable.Empty<Source>();
+                }
 
-
+                await _context.Sources.AddAsync(source);
+                await _context.SaveChangesAsync();
+                return _context.Sources;
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Error occurred: {ex.Message}");
+                throw;
+            }
+        }
+        //find row data by id 
+        [HttpGet]
+        public async Task<IEnumerable<Source>> GetDeleteNameSource(int id)
+        {
+            try
+            {
+                return await _context.Sources.Where(r => r.Id == id).ToListAsync();
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Error occurred: {ex.Message}");
+                throw;
+            }
+        }
+        //delete name roles
+        [HttpPost]
+        public async Task<IEnumerable<Source>> DeleteNameSource(int id)
+        {
+            try
+            {
+                var data = await _context.Sources.FindAsync(id);
+                if (data == null)
+                {
+                    return Enumerable.Empty<Source>();
+                }
+                _context.Sources.Remove(data);
+                _context.SaveChanges();
+                return _context.Sources;
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Error occurred: {ex.Message}");
+                throw;
+            }
+        }
 
     }
 }
